Make `segmentSizeMultiplier` a float.

+++ Models/Config.cs
@@ -9,7 +9,7 @@
         public bool  infiniteStamina                = false;
         public bool  infiniteAmmo                   = false;
         public float inventorySizeMultiplier        = 1f;
-        public int   segmentSizeMultiplier          = 1;
+        public float segmentSizeMultiplier          = 1f;
         public bool  disableAimSway                 = false;
         public bool  hideHelpIconNearCrosshair      = false;
         public float oreAndIngotStackSizeMultiplier = 1f;
